--- .env.example
@@ -1,7 +1,0 @@
-NETWORK_TESTNET_URL=https://abcde
-NETWORK_TESTNET_PRIVATE_KEY=0xasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasda
-NETWORK_MAINNET_URL=https://abcde
-NETWORK_MAINNET_PRIVATE_KEY=0xnasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasda
-
-BLOCK_EXPLORER_API_KEY=
-GAS_REPORTER_COIN_MARKET_CAP_API_KEY=
