+++ config/CollectionConfig.ts
@@ -6,10 +6,10 @@ const CollectionConfig: CollectionConfigInterface = {
     testnet: Networks.niskala,
     mainnet: Networks.arbitrumOne,
     contractName: "DataSharing",
-    platformAddress: "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266", // account test
+    platformAddress: "0x4c406d5fF749D73d0034be044b1C8DD57cF1134c", // account test
     domainEip712: "DataSharing",
     versionDomain: "1",
-    contractAddress: "0x1E8E93ff202e046D46E2d7fe9b0a8471F07e945D",
+    contractAddress: "0x922872cA2B2FC36cE54EF998Ce1532D774A6511E",
     //marketplaceIdentifier: "market-place-identifier",
     //marketplaceConfig: Marketpalce.openSea
 };
