--- scripts/5_deploy_test_token_native.ts
@@ -1,22 +1,0 @@
-import { ethers } from "hardhat";
-import CollectionConfig from "../config/CollectionConfig";
-import { NftContractType } from "../lib/NftContractProvider";
-import ContractArguments from "../config/ContractArguments";
-
-async function main() {
-
-  console.log("Deploying contract..");
-
-  // We get the contract to deploy
-  const Contract = await ethers.getContractFactory("TestToken");
-  const contract = await Contract.deploy() as unknown as NftContractType;
-
-  await contract.deployed();
-
-  console.log("Greeter deployed to:", contract.address);
-}
-
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
